Added Addressing response routing

+++ receivers/native/generic-core-2/src/main/c++/zcode/ZcodeBusInterrupt.hpp
@@ -39,7 +39,7 @@ public:
         return info.bus;
     }
 
-    uint8_t getFoundAddress() {
+    uint16_t getFoundAddress() {
         return address.address;
     }
